requirements.txt
asgiref==3.8.1
coverage==7.8.0
Django==5.2
django-rest==0.8.7
django-rest-framework==0.1.0
djangorestframework==3.16.0
drf-yasg==1.21.10
inflection==0.5.1
iniconfig==2.1.0
packaging==25.0
pillow==11.2.1
pluggy==1.5.0
psycopg2-binary==2.9.10
pytest==8.3.5
pytest-cov==6.1.1
pytest-django==4.11.1
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
ruff==0.11.5
six==1.17.0
sqlparse==0.5.3
uritemplate==4.1.1
